Invitations controller: now not needed to be logged in

+++ app/controllers/invitations_controller.rb
@@ -4,12 +4,17 @@ class InvitationsController < ApplicationController
     invitee = Traveller.find_by_invitation_url(params[:id])
     render_404 and return unless invitee
 
-    current_user.email = invitee.email
-    current_user.trips << invitee.trips
-    current_user.save
+    if current_user
+      current_user.email = invitee.email
+      current_user.trips << invitee.trips
+      current_user.save
 
-    invitee.destroy
+      invitee.destroy
+      redirect_to trip_destination_path(current_user.trips.last), flash: { success: "Invitation received successfully." }
+    else
+      session[:invitation_token] = params[:id]
+      redirect_to root_url, flash: { notice: "Please sign in or sign up to access the trips content" }
+    end
 
-    redirect_to trip_destination_path
   end
 end
